Capture GraphQL errors when filtering recordings
Pairs with https://github.com/replayio/backend/pull/6107 to make sure
that we will can display GraphQL errors that we get back when filtering
recordings. Originally I had a crappy UI in here for showing the actual
message, but it turns out that we've decided that for now, a broken JSON
search phrase resulting in no results is fine. All the same, I left the
logic in for capturing the error, because I could see it being useful in
the future.